Move sub-component specific methods where they belong
