+++ spec/components/provider_interface/application_choice_header_component_spec.rb
@@ -152,87 +152,12 @@ RSpec.describe ProviderInterface::ApplicationChoiceHeaderComponent do
     end
   end
 
-  describe '#deferred_offer_wizard_applicable?' do
-    it 'is true for a deferred offer belonging to the previous recruitment cycle' do
-      application_choice = instance_double(ApplicationChoice, status: 'offer_deferred')
-      allow(application_choice).to receive(:recruitment_cycle).and_return(RecruitmentCycle.previous_year)
-
-      expect(described_class.new(application_choice: application_choice, provider_can_respond: true).deferred_offer_wizard_applicable?).to be true
-    end
-
-    it 'is false if the provider cannot respond to the application' do
-      application_choice = instance_double(ApplicationChoice, status: 'offer_deferred')
-      allow(application_choice).to receive(:recruitment_cycle).and_return(RecruitmentCycle.previous_year)
-
-      expect(described_class.new(application_choice: application_choice, provider_can_respond: false).deferred_offer_wizard_applicable?).to be false
-    end
-
-    it 'is false when the application status is not deferred' do
-      application_choice = instance_double(ApplicationChoice, status: 'withdrawn')
-      allow(application_choice).to receive(:recruitment_cycle).and_return(RecruitmentCycle.previous_year)
-
-      expect(described_class.new(application_choice: application_choice, provider_can_respond: true).deferred_offer_wizard_applicable?).to be false
-    end
-
-    it 'is false when the application recruitment cycle is current' do
-      application_choice = instance_double(ApplicationChoice, status: 'offer_deferred')
-      allow(application_choice).to receive(:recruitment_cycle).and_return(RecruitmentCycle.current_year)
-
-      expect(described_class.new(application_choice: application_choice, provider_can_respond: true).deferred_offer_wizard_applicable?).to be false
-    end
-  end
-
-  describe '#deferred_offer_but_cannot_respond?' do
-    it 'is true if the provider cannot respond to the application' do
-      application_choice = instance_double(ApplicationChoice, status: 'offer_deferred')
-      allow(application_choice).to receive(:recruitment_cycle).and_return(RecruitmentCycle.previous_year)
-
-      expect(described_class.new(application_choice: application_choice, provider_can_respond: false).deferred_offer_but_cannot_respond?).to be true
-    end
-  end
-
-  describe '#deferred_offer_in_current_cycle?' do
-    it 'is true for a deferred offer without an open offered option' do
-      course_option = instance_double(CourseOption, course: instance_double(Course, open_on_apply: false))
-      application_choice = instance_double(ApplicationChoice, status: 'offer_deferred', recruitment_cycle: RecruitmentCycle.current_year)
-      allow(course_option).to receive(:in_next_cycle).and_return(false)
-      allow(application_choice).to receive(:current_course_option).and_return(course_option)
-
-      expect(described_class.new(application_choice: application_choice).deferred_offer_in_current_cycle?).to be true
-    end
-
-    it 'is false for a deferred offer with an open offered option' do
-      course_option = instance_double(CourseOption, course: instance_double(Course, open_on_apply: false))
-      application_choice = instance_double(ApplicationChoice, status: 'offer_deferred', recruitment_cycle: RecruitmentCycle.current_year)
-      allow(course_option).to receive(:in_next_cycle).and_return(course_option)
-      allow(application_choice).to receive(:current_course_option).and_return(course_option)
-
-      expect(described_class.new(application_choice: application_choice).deferred_offer_in_current_cycle?).to be false
-    end
-
-    it 'is false for a deferred offer from the previous cycle' do
-      course_option = instance_double(CourseOption, course: instance_double(Course, open_on_apply: false))
-      application_choice = instance_double(ApplicationChoice, status: 'offer_deferred', recruitment_cycle: RecruitmentCycle.previous_year)
-      allow(course_option).to receive(:in_next_cycle).and_return(course_option)
-      allow(application_choice).to receive(:current_course_option).and_return(course_option)
-
-      expect(described_class.new(application_choice: application_choice).deferred_offer_in_current_cycle?).to be false
-    end
-  end
-
   describe '#rejection_reason_required' do
     it 'is true for a rejected by default application without a rejection reason' do
       application_choice = instance_double(ApplicationChoice, status: 'rejected', rejected_by_default: true, rejection_reason: nil, structured_rejection_reasons: nil)
       allow(application_choice).to receive(:no_feedback?).and_return(true)
 
       expect(described_class.new(application_choice: application_choice, provider_can_respond: true).rejection_reason_required?).to be true
-    end
-
-    it 'is false if the provider cannot respond to the application' do
-      application_choice = instance_double(ApplicationChoice, status: 'rejected', rejected_by_default: true, rejection_reason: nil, structured_rejection_reasons: nil)
-      allow(application_choice).to receive(:no_feedback?).and_return(true)
-
-      expect(described_class.new(application_choice: application_choice).rejection_reason_required?).to be false
     end
 
     it 'is false for a rejected by default application with a rejection reason' do
@@ -252,49 +177,6 @@ RSpec.describe ProviderInterface::ApplicationChoiceHeaderComponent do
       application_choice = instance_double(ApplicationChoice, status: 'offer_deferred')
 
       expect(described_class.new(application_choice: application_choice, provider_can_respond: true).rejection_reason_required?).to be false
-    end
-  end
-
-  describe '#decline_by_default_text' do
-    it 'returns nil if the application is not in the offer state' do
-      application_choice = build_stubbed(:application_choice, status: 'awaiting_provider_decision')
-
-      expect(described_class.new(application_choice: application_choice).decline_by_default_text).to be_nil
-    end
-
-    describe 'returns the correct text when' do
-      it 'the dbd is today' do
-        application_choice = build_stubbed(
-          :application_choice,
-          status: 'offer',
-          decline_by_default_at: Time.zone.now.end_of_day,
-        )
-
-        expected_text = "at the end of today (#{application_choice.decline_by_default_at.to_fs(:govuk_date_and_time)})"
-        expect(described_class.new(application_choice: application_choice).decline_by_default_text).to eq(expected_text)
-      end
-
-      it 'the dbd is tomorrow' do
-        application_choice = build_stubbed(
-          :application_choice,
-          status: 'offer',
-          decline_by_default_at: 1.day.from_now.end_of_day,
-        )
-
-        expected_text = "at the end of tomorrow (#{application_choice.decline_by_default_at.to_fs(:govuk_date_and_time)})"
-        expect(described_class.new(application_choice: application_choice).decline_by_default_text).to eq(expected_text)
-      end
-
-      it 'the dbd is after tomorrow' do
-        application_choice = build_stubbed(
-          :application_choice,
-          status: 'offer',
-          decline_by_default_at: 3.days.from_now.end_of_day,
-        )
-
-        expected_text = "in 3 days (#{application_choice.decline_by_default_at.to_fs(:govuk_date_and_time)})"
-        expect(described_class.new(application_choice: application_choice).decline_by_default_text).to eq(expected_text)
-      end
     end
   end
 end
